requirements.txt
openai-whisper
moviepy
anthropic
librosa
numpy
soundfile
ffmpeg-python

